Crate embassy_hal_internal

embassy-hal-internal

An Embassy project.

Internal implementation details for Embassy HALs. DO NOT USE DIRECTLY. Embassy HALs (embassy-nrf, embassy-stm32, embassy-rp) already reexport everything you need to use them effectively.

Modules

atomic_ring_buffer

Atomic reusable ringbuffer.

drop

Types for controlling when drop is invoked.

interrupt

Interrupt handling for cortex-m devices.

ratio

Types for dealing with rational numbers.

Macros

impl_peripheral

Implement the peripheral trait.

interrupt_mod

Generate a standard mod interrupt for a HAL.

into_ref

Convenience converting into reference.

peripherals

Defining peripheral type.

peripherals_definition

Types for the peripheral singletons.

peripherals_struct

Define the peripherals struct.

Structs

PeripheralRef

An exclusive reference to a peripheral.

Traits

Peripheral

Trait for any type that can be used as a peripheral of type P.