Trait embassy_hal_internal::interrupt::InterruptExt

pub unsafe trait InterruptExt: InterruptNumber + Copy {
    // Provided methods
    unsafe fn enable(self) { ... }
    fn disable(self) { ... }
    fn is_active(self) -> bool { ... }
    fn is_enabled(self) -> bool { ... }
    fn is_pending(self) -> bool { ... }
    fn pend(self) { ... }
    fn unpend(self) { ... }
    fn get_priority(self) -> Priority { ... }
    fn set_priority(self, prio: Priority) { ... }
    fn set_priority_with_cs(self, _cs: CriticalSection<'_>, prio: Priority) { ... }
}

Represents an interrupt type that can be configured by embassy to handle interrupts.

Provided Methods

unsafe fn enable(self)

Enable the interrupt.

fn disable(self)

Disable the interrupt.

fn is_active(self) -> bool

Check if interrupt is being handled.

fn is_enabled(self) -> bool

Check if interrupt is enabled.

fn is_pending(self) -> bool

Check if interrupt is pending.

fn pend(self)

Set interrupt pending.

fn unpend(self)

Unset interrupt pending.

fn get_priority(self) -> Priority

Get the priority of the interrupt.

fn set_priority(self, prio: Priority)

Set the interrupt priority.

fn set_priority_with_cs(self, _cs: CriticalSection<'_>, prio: Priority)

Set the interrupt priority with an already-acquired critical section

Equivalent to set_priority, except you pass a CriticalSection to prove you’ve already acquired a critical section. This prevents acquiring another one, which saves code size.

Object Safety

This trait is not object safe.