embassy_hal_internal
Struct PeripheralRef

pub struct PeripheralRef<'a, T> { /* private fields */ }

An exclusive reference to a peripheral.

This is functionally the same as a &'a mut T. Thereโs a few advantages in having a dedicated struct instead:

Memory efficiency: Peripheral singletons are typically either zero-sized (for concrete peripherals like PA9 or SPI4) or very small (for example AnyPin, which is 1 byte). However &mut T is always 4 bytes for 32-bit targets, even if T is zero-sized. PeripheralRef stores a copy of T instead, so itโs the same size.

Code size efficiency. If the user uses the same driver with both SPI4 and &mut SPI4, the driver code would be monomorphized two times. With PeripheralRef, the driver is generic over a lifetime only. SPI4 becomes PeripheralRef<'static, SPI4>, and &mut SPI4 becomes PeripheralRef<'a, SPI4>. Lifetimes donโt cause monomorphization.

Implementations

impl<'a, T> PeripheralRef<'a, T>

pub fn new(inner: T) -> Self

Create a new reference to a peripheral.

pub unsafe fn clone_unchecked(&self) -> PeripheralRef<'a, T>
where T: Peripheral<P = T>,

Unsafely clone (duplicate) a peripheral singleton.

Safety

This returns an owned clone of the peripheral. You must manually ensure only one copy of the peripheral is in use at a time. For example, donโt create two SPI drivers on SPI1, because they will โfightโ each other.

You should strongly prefer using reborrow() instead. It returns a PeripheralRef that borrows self, which allows the borrow checker to enforce this at compile time.

pub fn reborrow(&mut self) -> PeripheralRef<'_, T>
where T: Peripheral<P = T>,

Reborrow into a โchildโ PeripheralRef.

self will stay borrowed until the child PeripheralRef is dropped.

pub fn map_into<U>(self) -> PeripheralRef<'a, U>
where T: Into<U>,

Map the inner peripheral using Into.

This converts from PeripheralRef<'a, T> to PeripheralRef<'a, U>, using an Into impl to convert from T to U.

For example, this can be useful to degrade GPIO pins: converting from PeripheralRef<โa, PB11>toPeripheralRef<โa, AnyPin>`.